requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
flask
faiss

-e .